+++ plugin/src/main/java/io/jenkins/plugins/forensics/git/miner/CommitCollector.java
@@ -18,15 +18,14 @@ import one.util.streamex.StreamEx;
 
 /**
  * Collects all commits for a git repository up to a given commit ID. The collected commits will be sorted ascending,
- * i.e. the list starts with the the given commit ID up to the current HEAD.
+ * i.e. the list starts with the given commit ID up to the current HEAD.
  *
  * @author Giulia Del Bravo
  * @author Ullrich Hafner
  */
 class CommitCollector {
     List<RevCommit> findAllCommits(final Repository repository, final Git git, final String latestCommitId,
-            final FilteredLog logger)
-            throws IOException, GitAPIException {
+            final FilteredLog logger) throws IOException, GitAPIException {
         ObjectId head = repository.resolve(Constants.HEAD);
         if (head == null) {
             logger.logError("No HEAD commit found in " + repository);
