+++ plugin/src/main/java/io/jenkins/plugins/forensics/git/util/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Provides utility classes to access Git repositories and to render Git commit hashes.
+ */
+@DefaultAnnotation(NonNull.class)
+package io.jenkins.plugins.forensics.git.util;
+
+import edu.umd.cs.findbugs.annotations.DefaultAnnotation;
+import edu.umd.cs.findbugs.annotations.NonNull;
